+++ src/test/java/test/org/fugerit/java/tool/etp/TestExcelToProps.java
@@ -4,7 +4,12 @@ import io.quarkus.test.junit.QuarkusTest;
 import org.fugerit.java.core.function.SafeFunction;
 import org.fugerit.java.core.io.StreamIO;
 import org.fugerit.java.core.lang.helpers.ClassHelper;
+import org.fugerit.java.tool.etp.ETPOutput;
+import org.fugerit.java.tool.etp.ExcelToPropsRest;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import java.io.InputStream;
 
 import static io.restassured.RestAssured.given;
 
@@ -26,6 +31,16 @@ class TestExcelToProps {
                 .post( "/fj-quarkus-tool/api/excel_to_props/convert/0/0/0/1" )
                 .then()
                 .statusCode(200);
+    }
+
+    @Test
+    void testConvertFacade() {
+        SafeFunction.apply( () -> {
+            try (InputStream bis = ClassHelper.loadFromDefaultClassLoader( "etp/test.xlsx" ) ) {
+                ETPOutput output = ExcelToPropsRest.convertWorker( bis, 0, 0, 1, 1 );
+                Assertions.assertNotNull( output );
+            }
+        } );
     }
 
 }