+++ src/test/java/test/org/fugerit/java/tool/ptj/TestPropsToJsonRest.java
@@ -8,6 +8,7 @@ import org.fugerit.java.core.io.StreamIO;
 import org.fugerit.java.core.lang.helpers.ClassHelper;
 import org.junit.jupiter.api.Test;
 
+
 import static io.restassured.RestAssured.given;
 
 @Slf4j